+++ NEWS.md
@@ -2,5 +2,7 @@
 
 * add experimental `find_files()` 
 
+* add highlighting of matches in source marker results
+
 * Cleaned up code for handling multiple directories in `sif()` and provide more
   helpful error message if directories do not exist