+++ NEWS.md
@@ -1,8 +1,12 @@
+# sifr 0.0.3.9000
+
+* add `open_files()`
+* add `sifr.use_markers` option to control default printing behaviour of `sif()`, 
+  `find_files()`, etc...
+
 # sifr 0.0.3
 
 * add experimental `find_files()` 
-
 * add highlighting of matches in source marker results
-
 * Cleaned up code for handling multiple directories in `sif()` and provide more
   helpful error message if directories do not exist